[feat] 마이페이지 프로필 헤더뷰로 변경
